Fix broken use of event waiter
It is only safe to use the event waiter for wake-up when the queue is
empty. If the queue has multiple items in it, we should drain them in
one shot. Otherwise we run the risk of getting two items in the queue
at the same time and the event only firing once.