Copy message buffer to fix concurrent modification
If either the broker's clients or its upstream produces messages fast
enough, they can trigger an out-of-order send because the to-be-sent
data isn't immediately copied out of the receiving buffer. It could sit
on the event queue briefly during which it would be overwritten with a
later message.

This is a WIP fix while I figure out a way to better structure memory
use to avoid these kinds of defensive copies.